pyproject.toml
[tool.poetry]
name = "dutycalendar"
version = "0.1.0"
description = ""
authors = ["Joris <jorisbeets@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
ics = "^0.7.2"
python-dateutil = "^2.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
